+++ packages/topics/package.json
@@ -36,9 +36,9 @@
   },
   "homepage": "https://github.com/newsuk/times-components#readme",
   "devDependencies": {
-    "@storybook/react-native": "3.4.1",
     "@times-components/eslint-config-thetimes": "0.4.1",
     "@times-components/jest-configurator": "0.6.2",
+    "@times-components/storybook": "0.9.3",
     "@times-components/tealium-utils": "0.2.5",
     "babel-core": "6.26.0",
     "babel-plugin-add-react-displayname": "0.0.5",
